# docs/openapi.yaml
openapi: 3.0.0
info:
  title: Plataforma de Auditoria Fiscal Preventiva (CFDI Intelligence)
  version: 1.0.0
  description: |
    API REST para la plataforma de deteccion automatica de riesgos fiscales.
    Permite gestionar empresas, ingestar CFDIs y estados de cuenta bancarios,
    conciliar movimientos, detectar riesgos fiscales, calcular scoring de
    cumplimiento y descargar CFDIs directamente del SAT via FIEL (e.firma).

    **Autenticacion**: Todos los endpoints (excepto /auth/register, /auth/login
    y /health) requieren un Bearer Token JWT obtenido al autenticarse.

    **Vencimiento del token**: 8 horas desde su emision.
servers:
  - url: https://api.cfdi-intelligence.com
    description: Produccion
  - url: http://localhost:8000
    description: Desarrollo local
...
security:
  - BearerAuth: []
paths:
...
  /api/v1/constancia/parsear:
    post:
      operationId: parsearConstancia
      tags: [Constancia]
      summary: Extraer datos de Constancia de Situacion Fiscal
      description: |
        Recibe un archivo PDF de la Constancia de Situacion Fiscal emitida por el SAT
        y extrae los datos fiscales estructurados: RFC, razon social, regimen fiscal,
        codigo postal, CURP, obligaciones, etc.

        El archivo PDF es guardado en el servidor y se retorna su nombre para
        referencia posterior al crear la empresa.
      security: []
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required: [archivo]
              properties:
                archivo:
                  type: string
                  format: binary
                  description: Archivo PDF de la Constancia de Situacion Fiscal del SAT
      responses:
        '200':
          description: Datos fiscales extraidos del PDF
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ConstanciaParseada'
              example:
                rfc: "XAXX010101000"
                razon_social: "Empresa Ejemplo SA de CV"
                regimen_fiscal: "601"
                cp_fiscal: "06600"
                curp: null
                obligaciones: ["IVA mensual", "ISR mensual"]
                representante_legal: "Juan Perez Lopez"
                rfc_representante: "PELJ800101ABC"
                constancia_path: "3fa85f64-5717-4562-b3fc-2c963f66afa6.pdf"
        '400':
          description: El archivo no es PDF
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                detail: "El archivo debe ser PDF"
        '422':
          $ref: '#/components/responses/UnprocessableError'
        '500':
          $ref: '#/components/responses/InternalError'

  # ============================================================
  # EMPRESAS
  # ============================================================
  /api/v1/empresas:
    get:
      operationId: listarEmpresas
      tags: [Empresas]
      summary: Listar empresas del contador
      description: |
        Retorna todas las empresas activas vinculadas al contador autenticado,
        ordenadas por fecha de vinculacion ascendente.
      responses:
        '200':
          description: Lista de empresas
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Empresa'
              example:
                - id: "a1b2c3d4-0001-0001-0001-000000000001"
                  rfc: "XAXX010101000"
                  razon_social: "Empresa Ejemplo SA de CV"
                  regimen_fiscal: "601"
                  cp_fiscal: "06600"
                  activo: true
        '401':
          $ref: '#/components/responses/UnauthorizedError'
  /api/v1/mis-empresas:
    post:
      operationId: agregarEmpresa
      tags: [Empresas]
      summary: Agregar empresa al contador
      description: |
        Crea una nueva empresa (o reutiliza una existente por RFC) y la vincula
        al contador autenticado. La operacion es idempotente: si la empresa ya
        esta vinculada, no se crea un duplicado.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AgregarEmpresaRequest'
            example:
              rfc: "XAXX010101000"
              razon_social: "Empresa Ejemplo SA de CV"
              regimen_fiscal: "601"
              cp_fiscal: "06600"
              curp: null
              obligaciones: ["IVA mensual", "ISR mensual"]
              representante_legal: "Juan Perez Lopez"
              rfc_representante: "PELJ800101ABC"
      responses:
        '201':
          description: Empresa vinculada correctamente
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EmpresaVinculadaResponse'
              example:
                mensaje: "Empresa vinculada correctamente"
                empresa_id: "a1b2c3d4-0001-0001-0001-000000000001"
                rfc: "XAXX010101000"
                razon_social: "Empresa Ejemplo SA de CV"
        '401':
          $ref: '#/components/responses/UnauthorizedError'
  /api/v1/empresas/{empresa_id}:
    get:
      operationId: obtenerEmpresa
      tags: [Empresas]
      summary: Obtener detalle de una empresa
      description: |
        Retorna todos los datos de una empresa especifica. El usuario autenticado
        debe tener acceso a la empresa.
      parameters:
        - $ref: '#/components/parameters/EmpresaId'
      responses:
        '200':
          description: Datos de la empresa
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Empresa'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '403':
          $ref: '#/components/responses/ForbiddenError'
        '404':
          $ref: '#/components/responses/NotFoundError'
  /api/v1/empresas/{empresa_id}/impuestos:
    patch:
      operationId: actualizarImpuestos
      tags: [Empresas]
      summary: Actualizar impuestos a declarar
      description: |
        Actualiza la lista de impuestos que la empresa debe declarar.
        Valores validos: iva, isr, ieps, ret_iva, ret_isr, diot.
      parameters:
        - $ref: '#/components/parameters/EmpresaId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ImpuestosRequest'
            example:
              impuestos: ["iva", "isr", "ret_iva"]
      responses:
        '200':
          description: Impuestos actualizados
          content:
            application/json:
              schema:
                type: object
                properties:
                  ok:
                    type: boolean
                  impuestos:
                    type: array
                    items:
                      type: string
              example:
                ok: true
                impuestos: ["iva", "isr", "ret_iva"]
        '400':
          $ref: '#/components/responses/BadRequestError'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '403':
          $ref: '#/components/responses/ForbiddenError'
        '404':
          $ref: '#/components/responses/NotFoundError'

  # ============================================================
  # INGESTA
  # ============================================================
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: |
        Token JWT obtenido al registrarse o hacer login.
        Incluir en el header: Authorization: Bearer {token}
        Duracion: 8 horas.
  parameters:
    EmpresaId:
      name: empresa_id
      in: path
      required: true
      schema:
        type: string
        format: uuid
      description: ID unico de la empresa (UUID)
      example: "a1b2c3d4-0001-0001-0001-000000000001"
  responses:
    UnauthorizedError:
      description: Token JWT ausente, invalido o expirado
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            detail: "Se requiere autenticacion"
    ForbiddenError:
      description: El usuario no tiene acceso a este recurso
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            detail: "Sin acceso a esta empresa"
    NotFoundError:
      description: El recurso solicitado no existe
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            detail: "Empresa no encontrada"
    BadRequestError:
      description: La solicitud contiene datos invalidos
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            detail: "Sin campos para actualizar"
...
    UnprocessableError:
      description: La entidad no pudo procesarse (datos semanticamente invalidos)
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            detail: "FIEL invalida: certificado corrupto"
    InternalError:
      description: Error interno del servidor
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            detail: "Error inesperado en el servidor"
...
  schemas:
...
    ErrorResponse:
      type: object
      required: [detail]
      properties:
        detail:
          type: string
          description: Descripcion del error
...
    Empresa:
      type: object
      properties:
        id:
          type: string
          format: uuid
        rfc:
          type: string
          example: "XAXX010101000"
        razon_social:
          type: string
        regimen_fiscal:
          type: string
          nullable: true
        cp_fiscal:
          type: string
          nullable: true
          description: Codigo postal del domicilio fiscal
        curp:
          type: string
          nullable: true
          description: CURP (solo personas fisicas)
        obligaciones:
          type: array
          nullable: true
          items:
            type: string
          description: Lista de obligaciones fiscales
        representante_legal:
          type: string
          nullable: true
        rfc_representante:
          type: string
          nullable: true
        impuestos_declarar:
          type: array
          nullable: true
          items:
            type: string
          description: Impuestos configurados para declarar
        activo:
          type: boolean
    AgregarEmpresaRequest:
      type: object
      required: [rfc, razon_social]
      properties:
        rfc:
          type: string
          description: RFC de la empresa (se convierte a mayusculas)
          example: "XAXX010101000"
        razon_social:
          type: string
        regimen_fiscal:
          type: string
          nullable: true
        cp_fiscal:
          type: string
          nullable: true
        curp:
          type: string
          nullable: true
        obligaciones:
          type: array
          nullable: true
          items:
            type: object
        representante_legal:
          type: string
          nullable: true
        rfc_representante:
          type: string
          nullable: true
    EmpresaVinculadaResponse:
      type: object
      properties:
        mensaje:
          type: string
        empresa_id:
          type: string
          format: uuid
        rfc:
          type: string
        razon_social:
          type: string
    ImpuestosRequest:
      type: object
      required: [impuestos]
      properties:
        impuestos:
          type: array
          items:
            type: string
            enum: [iva, isr, ieps, ret_iva, ret_isr, diot]
          description: Lista de claves de impuestos a declarar
    ConstanciaParseada:
      type: object
      properties:
        rfc:
          type: string
        razon_social:
          type: string
        regimen_fiscal:
          type: string
          nullable: true
        cp_fiscal:
          type: string
          nullable: true
        curp:
          type: string
          nullable: true
        obligaciones:
          type: array
          nullable: true
          items:
            type: string
        representante_legal:
          type: string
          nullable: true
        rfc_representante:
          type: string
          nullable: true
        constancia_path:
          type: string
          description: Nombre del archivo PDF guardado en el servidor
...